+++ hydra-node/src/Hydra/HeadLogic/State.hs
@@ -5,9 +5,10 @@
 module Hydra.HeadLogic.State where
 
 import Data.Map qualified as Map
-import Hydra.Chain (ChainStateType, HeadId, HeadParameters)
+import Hydra.Chain (ChainStateType, HeadParameters)
 import Hydra.ContestationPeriod (ContestationPeriod)
 import Hydra.Crypto (HydraKey, Signature, SigningKey)
+import Hydra.HeadId (HeadId)
 import Hydra.Ledger (ChainSlot, IsTx (..))
 import Hydra.Party (Party, deriveParty)
 import Hydra.Prelude