+++ hydra-node/test/Hydra/HeadLogicSnapshotSpec.hs
@@ -10,6 +10,7 @@ import Data.List qualified as List
 import Data.Map.Strict qualified as Map
 import Hydra.Chain (HeadParameters (..))
 import Hydra.Crypto (sign)
+import Hydra.HeadId (HeadId (..))
 import Hydra.HeadLogic (
   CoordinatedHeadState (..),
   Effect (..),
@@ -48,13 +49,14 @@ spec = do
                 , otherParties = List.delete party threeParties
                 , contestationPeriod = defaultContestationPeriod
                 }
+        testHeadId = HeadId "1234"
 
     let coordinatedHeadState =
           CoordinatedHeadState
             { localUTxO = initUTxO
             , allTxs = mempty
             , localTxs = mempty
-            , confirmedSnapshot = InitialSnapshot initUTxO
+            , confirmedSnapshot = InitialSnapshot testHeadId initUTxO
             , seenSnapshot = NoSeenSnapshot
             }
     let sendReqSn =
@@ -64,7 +66,7 @@ spec = do
                   NetworkEffect ReqSn{} -> True
                   _ -> False
               )
-    let snapshot1 = Snapshot 1 mempty []
+    let snapshot1 = Snapshot testHeadId 1 mempty []
 
     let ackFrom sk vk = NetworkEvent defaultTTL vk $ AckSn (sign sk snapshot1) 1
 
@@ -90,7 +92,7 @@ spec = do
 
       it "does NOT send ReqSn when we are the leader but snapshot in flight" $ do
         let tx = aValidTx 1
-            sn1 = Snapshot 1 initUTxO mempty :: Snapshot SimpleTx
+            sn1 = Snapshot testHeadId 1 initUTxO mempty :: Snapshot SimpleTx
             st = coordinatedHeadState{seenSnapshot = SeenSnapshot sn1 mempty}
             outcome = update (envFor aliceSk) simpleLedger (inOpenState' [alice, bob] st) $ NetworkEvent defaultTTL alice $ ReqTx tx