+++ hydra-test-utils/test/Test/HSpec/MarkdownFormatterSpec.hs
@@ -1,3 +1,5 @@
+{-# LANGUAGE DerivingStrategies #-}
+
 module Test.HSpec.MarkdownFormatterSpec where
 
 import Hydra.Prelude